refactor(client): HEADLESS-00 explicitly get client by type
