+++ src/graphql/utils.ts
@@ -1,0 +1,3 @@
+// Re-exporting types and utilities so consumers don't have to import from the @apollo/client package
+export type { MutationOptions, QueryOptions } from '@apollo/client';
+export { gql } from '@apollo/client';
