+++ src/pages/index.tsx
@@ -1,4 +1,3 @@
-import { gql } from '@apollo/client';
 import { GetServerSideProps } from 'next';
 import Head from 'next/head';
 import React from 'react';
@@ -12,7 +11,8 @@ import {
   ProductTilesConnection,
   query as ProductTilesQuery,
 } from '../components/ProductTiles';
-import { serverClient } from '../graphql/server';
+import { getServerClient } from '../graphql/server';
+import { gql } from '../graphql/utils';
 
 interface HomePageQuery {
   site: MergeDeep<
@@ -32,7 +32,9 @@ interface HomePageQuery {
 }
 
 export const getServerSideProps: GetServerSideProps = async () => {
-  const { data } = await serverClient.query<HomePageQuery>({
+  const client = getServerClient();
+
+  const { data } = await client.query<HomePageQuery>({
     query: gql`
       query HomePageQuery($pageSize: Int = 4) {
       site {
